Vue.js: 表示部分はコンポーネント化しました
