+++ app/vue/src/components/GetRecord.vue
@@ -27,19 +27,7 @@
 
     <div style="display:none" id="content" v-if="results">
       <div v-for="(result, key, index) in results" :key="index">
-        <div class="bg-white hover:bg-grey-light m-3 max-w-sm shadow-md rounded-md overflow-hidden" v-on:click="showRecord(result.id)">
-          <div class="text-left p-1 sm:text-left sm:flex-grow">
-            <span class="text-sm leading-tight">{{ result.title }}</span>
-            <span class="text-xs leading-tight text-grey-dark sm:inline">{{ result.subtitle }}</span>
-          </div>
-          <div class="p-1">
-            <strong class="text-xl leading-tight">¥{{ result.cost }}</strong>
-          </div>
-          <div class="text-left sm:text-left sm:flex-grow p-1">
-            <img src="../assets/icon_arrow_right.png" class="w-2"/>
-            <span class="text-xs text-grey-dark m-1">¥{{ result.compare }}({{ result.rate }}%)</span>
-          </div>
-        </div>
+        <row :result="result" />
       </div>
     </div>
   </div>
@@ -48,6 +36,7 @@
 <script>
 import axios from 'axios'
 import moment from 'moment'
+import BaseRow from './BaseRow'
 
 export default {
   name: 'GetRecord',
@@ -60,12 +49,15 @@ export default {
       end: new Date()
     }
   }),
+  components: {
+    'row': BaseRow,
+  },
   created () {
     // fetch the data when the view is created and the data is already being observed
-    this.getJson(this.$route.params.index)
+    this.getJson()
   },
   methods: {
-    getJson: function(id) {
+    getJson: function() {
       this.loading = true
 
       axios.post('/getReport')
@@ -79,18 +71,7 @@ export default {
              this.loading = false
              document.getElementById("content").style.display = "block";
            })
-    },
-    showRecord: function(id) {
-      this.$router.push({ name: 'ShowRecord', params: { index: id } });
     }
   }
 }
 </script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-.arrow {
-  width: 12px;
-  height: 12px;
-}
-</style>
