+++ app/vue/src/components/ShowRecord.vue
@@ -10,15 +10,7 @@
 
     <div style="display:none" id="content" v-if="results">
       <div v-for="(result, key, index) in results" :key="index">
-        <div class="bg-white hover:bg-grey-light m-3 max-w-sm shadow-md rounded-md overflow-hidden">
-          <div class="text-left p-1 sm:text-left sm:flex-grow">
-            <span class="text-sm leading-tight">{{ result.title }}</span>
-            <span class="text-xs leading-tight text-grey-dark sm:inline">{{ result.subtitle }}</span>
-          </div>
-          <div class="p-1">
-            <strong class="text-xl leading-tight">¥{{ result.cost.toFixed(2) }}</strong>
-          </div>
-        </div>
+        <DetailRow :result="result" />
       </div>
     </div>
   </div>
@@ -26,6 +18,7 @@
 
 <script>
 import axios from 'axios';
+import DetailRow from './DetailRow'
 
 export default {
   name: 'ShowRecord',
@@ -34,6 +27,9 @@ export default {
     results: null,
     error: null
   }),
+  components: {
+    DetailRow,
+  },
   created () {
     // fetch the data when the view is created and the data is already being observed
     this.getJson(this.$route.params.index)
